Add a small class to create a class instance on device.
And a unit test for it.  It is not just creating a device
object but a virtual class instance on device and then
invoking a virtual function through the base class on
device.

This is meant to be a simple example and the start of the
discussion on how to structure the code for future
development of GPU enabled Kokkos kernels.