Initial stk-ngp usage.
The two main classes in stk-ngp are ngp::Mesh and ngp::Field. These
classes are basically mechanisms for copying the corresponding stk::mesh
data to the gpu device, and then traversing or accessing the data on
the gpu device.

In the case where you're not using a gpu, then the ngp classes are
thin pass-throughs to regular stk::mesh objects. In other words, no
copy is made unless you are using a device with a distinct
memory space.