Remove nalu_stk::CommNeighbors and use stk::CommNeighbors.
nalu_stk::CommNeighbors came first as a prototyping exercise,
then the code moved into stk and continued to be improved
and fixed, including freeing the communicator that the
neighbor-communication produces. Getting rid of the prototype
and using the stk version should fix some out-of-communicators
errors that have been seen on some long-running simulations.