Master element on GPU unit test.
This takes a single function of a single master element,
  HexSCS::shape_fcn
and attempts to execute this function on the GPU.  This
is more of a prototype than a full implimentation.  I
think they call it a "straw man" that everyone can throw
rocks at and such.

Some highlights:

Added masterElementScs_ and masterElementScv_ to AlgTraits.
This is just a type name and allows for a function templated on
AlgTraits to have concrete master elements.  Avoids virtual
functions.

Added
template<typename AlgTraits>static typename
AlgTraits::masterElementScs_* get_surface_master_element();
and
template<typename AlgTraits>static typename
AlgTraits::masterElementScv_* get_volumee_master_element();
which create master elements on the GPU and then returns the
pointer on subsequent calls.