Finish off most of the gcc warnings.
Just rearrange some code to avoid warnings.

There was one fall-through case statement where I added
   __attribute__ ((fallthrough));
to prevent the warning. A Google search turned up several
ways to avoid the warning but this seemed the most
explicit but it might be gcc only.  C++17 has a [[fallthrough]]
once we move to C++17.

There are still warnings about floating point equality comparisons.
This is comparing to zero which I think is valid in this case and
did not want to change to a comparison to epsilon. I think we could
disable this one warning.