Prepare to move data from MasterElement to derived classes.
This will make it easier to use static data and Views which
then make it easier to use the master element classes on the GPU.

These are the standard get/set type functions for all the data
members marked as virtual. Once the data is moved to the derived
classes these will become pure virtual or throw if not implimented
in derived classes.