Add explicit error message regarding mesh_motion change
Pull request #77 introduced a backward incompatible change in the
implementation of mesh motion within Nalu-Wind codebase. This required a
change in the specification of mesh_motion witin the input file.
mesh_motion is now a section within the realm and not a subsection of
solution_options. Users using newer versions of nalu-wind executable are
presented with a cryptic error message:

```
terminate called after throwing an instance of 'std::runtime_error'
  what():  Only wind_energy and tornado user functions supported
```

This commit introduces two changes:

- Throws an error if it detects a mesh_motion section within
  solution_options, alerting the user to upgrade their input file

- Prints a warning if the wind_energy wall BC user function is provided
  in the input file.