Fix the intel seg-fault.
Still not completely clear on why it was seg-faulting. But it seems
related to the fact that MultiDimViews was packing the shared-mem-views
by value into flat arrays. So instead, and only if not using Cuda, we
now store them in separate pointers. I.e., an array of pointers. This
requires a new and delete operation, which we won't do on GPU.