Add a "using AlgTraits" to master elements.
As part of the master element refactor it was
decided to template the kernels on an AlgTraits
and then instantiate the master element based
on that.

There are two problems classes: Hex8FEM and the
higher order elements. It remains to be worked
out how these fit in the scheme of things.