Overset: provide node/cell volumes to TIOGA API for hole-cutting
Provide the dual nodal volume (node resolution) and element volume (cell
resolution) to TIOGA for use in hole-cutting process. The volumes computed in
Nalu-Wind use the MasterElement basis that is used for solution, and also will
handle the volumes of the shared nodes across processor interfaces correctly.
This reconciliation of cell volumes across processor interfaces is not easy in
TIOGA without support from the solver.