Change output behavior of Nalu-Wind
Currently, Nalu-Wind requires user to explicitly list the variables in the
output section. This commit changes the default behavior so that certain
variables are automatically added to the output list without the user explicitly
listing them in the output_variables list. The new list
`disabled_output_variables` is provided to explicitly disable output of these
variables.