TpetraSegregatedSystem: adding the new TpetraSegregatedLinearSystem class (see issue #249)
Create the new class and adds the necessary hooks in LinearSystem calss and CMakeList.txt
Running the code on dgNonConformalThreeBlade goes through the simulation but more check probabaly would be good to make sure that the solution is accurate enough.
A comparison with the segregated system using Hypre would be good too!