Transition terms in EquationSystems to their NGP versions (#300)
This pull-request continues the effort from #277 to replace edge/node algorithms/supplemental algorithms with their NGP counterparts (see #238 for details on the kernel transition status). With this change, we have all the necessary NGP assembly terms to be able to run an ABL precursor case with periodic, symmetry, and wall function terms.