Make get_entity_master more forgiving for ghost entities.
Added a flag so that get_entity_master can optionallly return an
invalid entity if master is not found, instead of throwing an exception.

In particular, when filling the entity_to_col_LID mapping-vector,
we traverse all of the entities (including ghosts) and occasionally
an entity won't have a local master. In those cases, this is a
location in the mapping-vector that won't be accessed anyway, so
we just set it to -1 and continue.