Revert "Revert "Set stack size on GPU.""
This reverts commit d4706afdc5c01aaa86853d82ba5dd1cb498bc2ac.

It was thought that the original commit broke the regression
tests but the cause was traced to something else.

Setting the stack size to something larger than the default
seems to be necessary but the size needed is hard to determine.
Too small and the runs seg fault, too large and the GPU will run
out of memory.