Step 2 of coeff-applier cleanup.
Make coeff-applier pointers be owned/managed by
LinearSystem. By default, get_coeff_applier returns a device pointer.
This also makes sense when running a non-cuda build, since the
host is the same as the device.
If we need a host pointer when running a cuda build, we can add a
get_host_coeff_applier.

Next step will be to try to move the extract-diagonal code into
something (a functor?) owned by SolverAlgorithm, and also put
the fix-overset code in that.