Fix segfaults on MacOS introduced by #432 and #434
Updates to Tpetra CoeffApplier code introduced segfaults on MacOS where the
ngp::Mesh instance stored within that object was deleted after
stk::mesh::BulkData was deleted. This commit simplies Tpetra CoeffApplier so
that it doesn't handle diagonal extraction anymore and just deals with sum_into
for the linear system.

This commit also restores the old behavior of SolverAlgorithm::apply_coeff for
all NGP assembly. For momentum solves, it will extract the diagonal term for use
with pressure Poisson solve and, for overset meshes, it restores the reset of
lhs/rhs coefficients for fringe and hole nodes.