GPU version of GradP open boundary algorithm.
First had to create a unit test for
AssembleNodalGradPOpenBoundaryAlgorithm to have a result
for before GPUing.

Then wrote the GPU version a couple of times before getting
that right and then more time to get the results of the
old and new unit tests to match.

The unit test might not be correct as the surface mesh part
seems to include too many faces.  But the old CPU and new GPU
versions of the algorithm give the same result which is something.

The unit tests pass but the other tests are failing for me for
what appears to be an unrelated problem having to do with matrix
assembly. Might need to update Trilinos.