Work-around to avoid deep_copy of scratch-views in unit-test.
Not sure why these deep_copy usages were problematic...
Replacing the deep_copy calls with a simple set_vals function is
workable only because the unit-tests in question weren't crossing
the boundary between cuda host and device. The copy-and-interleave
test was already ifdef'd to not be enabled for cuda, and the
'HO' test was purely non-cuda anyway.