Move to GPU versions of gradient kernels.
There was a problem with mixing GPU kernels
and CPU gradient contributions since the
sum into is then in different memory spaces.

This seems to fix much of the difference
between the CPU and GPU results for the
channel flow test we have been using to
debug this.

There are still differences in the results.
A deep dive on the contributions for a single
entry in the right hand side of the momentum
solve shows the order of operations can have
an impact since there are values summed in
of O(10^3) that almost cancel out in the final
result of O(10^-5).