Fix 2x discrepancy between CPU and GPU for TKE periodic BC node rhs coeffs.
Now, comparing log files for CPU vs GPU runs (for just 1 MPI rank
in each case, to remove that as a variable), the TKE residuals
match very well, and the 'Mean system norm' matches in 9 decimal
places after 10 time steps.

For follow-on work, we do need to convert to using the ngp version
of periodic_field_update which should help performance.