Two ngp fixes:
ElemDataRequestsGPU destructor needs to be KOKKOS_FUNCTION to
stop a compiler warning on gpu.

The matrix-free file needs the sync_to_device calls to fix
some diffs that appear for a coming stk update. (Stk is gradually
changing, during sierra milestone optimizations, to do less
implicit syncs so that requires that apps have to be more
careful to do syncs when they need them.)