Multi-realm overset capability (#641)
Adds a regression-test that demonstrates the ability to perform decoupled overset solves with linear system solves in two different realms. The regression test recreates the rotating cylinder regression test where the cylinder and background meshes are assigned to different realms. While this is the closest to the multi-solver overset capability that we can get to with the current nalu-wind executable and regression test infrastructure, the test exercises the nalu-wind side of the API that will be used for hybrid overset simulations. The golds were generated using GCC-7.4.0 on NREL Eagle, but likely would require re-blessing on the test machine.

To enable the same equation systems to be solved in multiple realms the LinearSolver creation and reinitialization interface had to be tweaked so that it took more than just the equation ID. Otherwise the linear solver instances from subsequent realms overwrote the ones from the preceding realm leading to segfaults in moving mesh simulations and memory leaks in non-moving mesh simulations.