Allow users to fix hub epsilon
The current nacelle model determines epsilon automatically
based on the paramaters that are supplied.
One side affect of this model is as cD and area are increased so is
epsilon, making the nacellel forcing more diffuse.
This change allows users to fix the hub epsilon to concentrate the
actuator force from the hub instead of having it be calculated
automatically for them.