Overset: Adjust resolutions for mandatory fringes (#730)
Adds an option to TIOGA overset connectivity logic that is used to provide
hints to TIOGA regarding nodes/cells that are adjacent to the overset boundaries
and must be ignored when considering potential donors. This is done by setting
a high value of nodal/cell resolutions, the criteria that TIOGA uses to determine
potential donors.