Fix bug in NGP loops unit tests for edges
Updates in PR #832 exposed a long-standing bug within the unit-test for the
basic edge loop. This test was not really executing the loops because
`create_edges` was never called. Updates in PR #832 changed the behavior of
creating hex8 meshes where edges are always created. Now the edge loops were
executing in the NGPLoops tests, and it exposed the fact that the device fields
were never synced to the host before the checks were performed.