Simpler HYPRE IJMatrix Assemble:
In this branch, the data structures for the assembled matrix (i,j,values) are combined into one large
monolithic buffer for owned and shared. The owned part comes first, then the shared. In addition,
the number of contributions (nnz, including repeats) on rank i from all other ranks is computed. This
allows us to resize the monolithic buffers to be nnz_owned + max(nnz_shared, nnz_shared_recv). Then,
this buffer can be reused inside IJMatrixAssemble thus saving a substantial amount of memory. Since this
optimized hypre assembly algorithnm is not yet in Hypre master, the API methods that indicate to
Hypre to use this new algorithm are disabled (though the code is left in for future use). The current
version of the Nalu data structure works with Hypre master (with the disabled code). As a part of this
development, the row_indices/counts has been replaced by rows which is the same size of as cols/values.
This allows us to use simpler IJMatrixSet/AddToValues2 APIs which use less memory and have fewer kernel
calls. We have added the option "simple_hypre_matrix_assemble", which will us to toggle between the
optimized and more general assembly algorithms, and will be useful once the optimized version is in
Hypre master.

Next, for all calls to sumInto* (HypreLinearSystem and HypreUVWLinearSystem), we have replaced the
binarySearch method for finding matrix write location with a sort and a constrained linear search. Because
the matrix locations defined by the column are sorted, one can use the old linear search as the starting
point for the next matrix write location, defined by the next column, which is sorted in ascending order.
These optimizations are facilitated by moving all read-only data structures in the numeric assembly to
RandomAccess Kokkos views. That is, we attempt to use fast Texture memory where possible.

We have added the ability to dump matrix statistics (rows, nnz per rank). This is useful for measuring
the effect of load balancing strategies and partition algorithms like parmetis. By default this is turned off
by default. Moroever, if toggled, we only do dump matrix stats once per equation system construction after
the global matrix assembly. The option is "dump_hypre_matrix_stats".