Add 'sync_host' call to avoid kokkos error-check.
This fixes a GPU error which happens in debug builds. It doesn't
seem to get hit in a release-mode build. But in a debug build kokkos
is complaining that the host-modify flag is being raised when the
device-modify flag is already raised. Calling sync_host before
modify_host makes it work.