IDDES updates:
There appears to be a bug in the Strelets upwinding.  I believe it has to do with some terms not being computed on the first timestep.
Possibly due to the order in which the algorithms are called.
However, since we haven't seen compelling evidence that we need the Strelets upwinding for our cases of interest, this PR
makes it so you have to explicitly turn it on when running IDDES with `strelets_upwinding: yes` in `solution_options`.
Other wise we use the default upwinding/Peclet blending methods in the code.
Users can only turn on Strelets upwinding with IDDES otherwise the parameter is ignored.

Additional fixes were required to get the sync/modified calls correct for IDDES specific fields