Fix the unit test LowMachKernelHex8Mesh.NGP_nodal_grad_popen
This has been failing on the dashboard because the values coming
back from the device algorithm were zero when checked by the
unit test.
The reason is because the field being modified wasn't being
marked as modified by the 'modify_on_device()' method, and then
the values were being lost by a sync_to_host/modify_on_host
sequence that occurs when the nodal-grad execute() returns.

It is possible that this is responsible for correctness issues in
regression tests, I'm not positive...