WIP: Remove some Fortran code. Issue #902
The Fortran code to compute determinants for the various
master elements was removed. These functions took double
as the datatype instead of SIMD types. The change is to
add functions that match the SIMD function calls but
using double so there is one consistent interface to
compute determinants through the master element classes.

There are many changes as the function signatures that
were changed are widely used and there is still some
work to do to clean up the code. And there is still
a lot of Fortran code to remove even after these changes.