Add unit-test fixture for cylinder mesh
Uses Neil Matula's code for starting with a generated mesh (which
is a brick-shaped domain of hex-8 elements) and transforms the
node coordinates to make the mesh into a cylinder.

The fixture is in UnitTestUtils.h for now (could be pulled out
to a separate header).

A simple test that uses the fixture to demonstrate creating
the fixture and doing a trivial manipulation of field-data values
on device and checking their correctness on host, is located
in UnitTestCylinderMesh.C