Work to merge NgpFieldManager into new field manager.
So, the idea is to get rid if nalu_ngp::FieldManager
and push the logic from that class into nalu::FieldManager.
One class that heavily uses nalu_ngp::FieldManager is
ElemDataRequestsGPU.  So template ElemDataRequestsGPU
on the FieldManager type so either can be used.
Then slowly change all of the occurrences of ElemDataRequestsGPU
from using nalu_ngp::FieldManager to using nalu::FieldManager.

Lots of cleanup of the interface and some of the implimentation details.

Clang format.