Example usage of FieldManager
This PR demonstrates how a Kernel can be converted to use a
FieldManager.

The main goal is to move the field registration away from higher level
abstractions like realm and equationsystem, to lower ones (algs and
kernels).

This will essentially eliminate the need for `RegisterNodalFields` type
calls, and will eliminate a range of bugs that can come in from not
registering fields in the right place.  Having kernels and algs register
the fields they need will also reduce the amount of fixturing needed for
unit tests since constructing the object will register the fields for
you.  You can then finalize the mesh and populate the fields with
initial values. (Not demonstrated in this PR).