Replace single stk::mesh::Part pointer with stk::mesh::PartVector. Issue #1183
PartVectors allow for the specification of a field on a list of parts
instead of one part at a time. For some classes it was only possible
to pass a single part pointer for the field specification where a
part vector would be more flexible.

Started with register_nodal_fields:
from: register_nodal_fields(stk::mesh::Part* part)
to:   register_nodal_fields(const stk::mesh::PartVector &part_vec)

The stk::mesh::put_field_on_mesh function does not take a PartVector
so create a Selector form a PartVector:
  stk::mesh::Selector selector = stk::mesh::selectUnion(part_vec);
and use that instead. There were a couple of places where the
construction of a PartVector from a single Part pointer:
   PartVector(1,Part*)
was useful.

There were a lot of files changed, but the changes are trivial.