Updating IBlank
I am not sure update shared is the right operation. It seems like if a cell is detected as fringe anywhere it should be fringe even if it is not the owner.  In this case `parallel_max` makes more sense to me.

This could explain the behavior https://github.com/Exawind/exawind-driver/issues/65